Create separate createVolumeTask helper func for windows
