+++ agent/engine/engine_integ_test.go
@@ -37,7 +37,6 @@ import (
 	"github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi"
 	"github.com/aws/amazon-ecs-agent/agent/dockerclient/sdkclientfactory"
 	"github.com/aws/amazon-ecs-agent/agent/engine/dockerstate"
-	taskresourcevolume "github.com/aws/amazon-ecs-agent/agent/taskresource/volume"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
@@ -416,64 +415,6 @@ func TestEngineSynchronize(t *testing.T) {
 
 	verifyContainerStoppedStateChange(t, taskEngine)
 	verifyTaskStoppedStateChange(t, taskEngine)
-}
-
-func TestSharedAutoprovisionVolume(t *testing.T) {
-	taskEngine, done, _ := setupWithDefaultConfig(t)
-	defer done()
-	stateChangeEvents := taskEngine.StateChangeEvents()
-	// Set the task clean up duration to speed up the test
-	taskEngine.(*DockerTaskEngine).cfg.TaskCleanupWaitDuration = 1 * time.Second
-
-	testTask, err := createVolumeTask(t, "shared", "TestSharedAutoprovisionVolume", "TestSharedAutoprovisionVolume", true)
-	require.NoError(t, err, "creating test task failed")
-
-	go taskEngine.AddTask(testTask)
-
-	verifyTaskIsRunning(stateChangeEvents, testTask)
-	verifyTaskIsStopped(stateChangeEvents, testTask)
-	assert.Equal(t, *testTask.Containers[0].GetKnownExitCode(), 0)
-	assert.Equal(t, testTask.ResourcesMapUnsafe["dockerVolume"][0].(*taskresourcevolume.VolumeResource).VolumeConfig.DockerVolumeName, "TestSharedAutoprovisionVolume", "task volume name is not the same as specified in task definition")
-	// Wait for task to be cleaned up
-	testTask.SetSentStatus(apitaskstatus.TaskStopped)
-	waitForTaskCleanup(t, taskEngine, testTask.Arn, 5)
-	client := taskEngine.(*DockerTaskEngine).client
-	response := client.InspectVolume(context.TODO(), "TestSharedAutoprovisionVolume", 1*time.Second)
-	assert.NoError(t, response.Error, "expect shared volume not removed")
-
-	cleanVolumes(testTask, taskEngine)
-}
-
-func TestSharedDoNotAutoprovisionVolume(t *testing.T) {
-	taskEngine, done, _ := setupWithDefaultConfig(t)
-	defer done()
-	stateChangeEvents := taskEngine.StateChangeEvents()
-	client := taskEngine.(*DockerTaskEngine).client
-	// Set the task clean up duration to speed up the test
-	taskEngine.(*DockerTaskEngine).cfg.TaskCleanupWaitDuration = 1 * time.Second
-
-	testTask, err := createVolumeTask(t, "shared", "TestSharedDoNotAutoprovisionVolume", "TestSharedDoNotAutoprovisionVolume", false)
-	require.NoError(t, err, "creating test task failed")
-
-	// creating volume to simulate previously provisioned volume
-	volumeConfig := testTask.Volumes[0].Volume.(*taskresourcevolume.DockerVolumeConfig)
-	volumeMetadata := client.CreateVolume(context.TODO(), "TestSharedDoNotAutoprovisionVolume",
-		volumeConfig.Driver, volumeConfig.DriverOpts, volumeConfig.Labels, 1*time.Minute)
-	require.NoError(t, volumeMetadata.Error)
-
-	go taskEngine.AddTask(testTask)
-
-	verifyTaskIsRunning(stateChangeEvents, testTask)
-	verifyTaskIsStopped(stateChangeEvents, testTask)
-	assert.Equal(t, *testTask.Containers[0].GetKnownExitCode(), 0)
-	assert.Len(t, testTask.ResourcesMapUnsafe["dockerVolume"], 0, "volume that has been provisioned does not require the agent to create it again")
-	// Wait for task to be cleaned up
-	testTask.SetSentStatus(apitaskstatus.TaskStopped)
-	waitForTaskCleanup(t, taskEngine, testTask.Arn, 5)
-	response := client.InspectVolume(context.TODO(), "TestSharedDoNotAutoprovisionVolume", 1*time.Second)
-	assert.NoError(t, response.Error, "expect shared volume not removed")
-
-	cleanVolumes(testTask, taskEngine)
 }
 
 func TestLabels(t *testing.T) {
